refactor: embeddings.rs
