+++ core/Cargo.toml
@@ -47,6 +47,7 @@ version.workspace = true
 # Enable vendored OpenSSL for aarch64 cross-compilation
 # Cross-compilation containers lack aarch64-specific OpenSSL dev packages
 [target.aarch64-unknown-linux-gnu]
+
 [target.aarch64-unknown-linux-gnu.dependencies]
 openssl = {version = "0.10", features = ["vendored"]}
