+++ docs/connector/redis_vector_search.md
